+++ lib/src/proste_custom_dialog.dart
@@ -43,13 +43,13 @@ class ProsteCustomDialog extends StatefulWidget {
   final Widget? header;
 
   /// 标题组件
-  final Text? title;
+  final Widget? title;
 
   /// 标题的padding
   final EdgeInsets titlePadding;
 
   /// 描述文字
-  final Text content;
+  final Widget content;
 
   /// 描述的间距
   final EdgeInsets contentPadding;